requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-genai==0.8.0
pydantic==2.10.0
requests==2.31.0
gunicorn==22.0.0
